refactor: extract AppErrorTypes.swift from AppErrorPresenter.swift
Extracts 3 types (AppErrorOperation enum, AppErrorNormalization struct,
AppErrorPresentationResult struct) from AppErrorPresenter.swift into a new
AppErrorTypes.swift.

Byte-identical move; SHA-256 e37ff7b83667ba6a2cc943ca7abda7fa4485f8a8b2a365ec0f5672f81f94decc
verified against origin/main:AppErrorPresenter.swift lines 3-27.

Reduces AppErrorPresenter.swift from 246 to 221 lines.

Tests: 667 (unchanged).
